Changed an int size to avoid a size mismatch on 64-bit systems.
git-svn-id: svn://svn.gna.org/svn/mysqlpp/trunk@1025 5946f024-35f4-0310-b1b3-9a3be3380cfb

+++ lib/query.cpp
@@ -151,7 +151,7 @@ void Query::parse()
 {
 	std::string str = "";
 	char num[4];
-	long int n;
+	int n;
 	char option;
 	std::string name;
 	char *s, *s0;
